Process People_Tested and People_Hospitalized from new datasets
The fields are saved as 'tested' and 'hospitalized' for simplicity in the
database documents. This fields are available for US country_province's.
The issue is they don't contain information about the county/administration.

As a result they (tested, hospitalized) have a value different than
zero in entries that have an Unassigned county/administration.

Signed-off-by: Marius Avram <marius.avram1309@gmail.com>